+++ src/OpenTelemetry.AutoInstrumentation/CallTarget/Handlers/Continuations/ContinuationGenerator.cs
@@ -15,7 +15,7 @@ internal class ContinuationGenerator<TTarget, TReturn>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     protected static TReturn ToTReturn<TFrom>(TFrom returnValue)
     {
-#if NET6_0_OR_GREATER
+#if NET8_0_OR_GREATER
         return Unsafe.As<TFrom, TReturn>(ref returnValue);
 #else
         return ContinuationsHelper.Convert<TFrom, TReturn>(returnValue);
@@ -25,7 +25,7 @@ internal class ContinuationGenerator<TTarget, TReturn>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     protected static TTo FromTReturn<TTo>(TReturn? returnValue)
     {
-#if NET6_0_OR_GREATER
+#if NET8_0_OR_GREATER
         return Unsafe.As<TReturn?, TTo>(ref returnValue);
 #else
         return ContinuationsHelper.Convert<TReturn?, TTo>(returnValue);