+++ test/IntegrationTests/Helpers/MockCorrelationCollector.cs
@@ -1,6 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
-#if NET6_0_OR_GREATER
+#if NET8_0_OR_GREATER
 using System.Collections.Concurrent;
 using System.Diagnostics;
 using FluentAssertions;