+++ test/IntegrationTests/PluginsTests.cs
@@ -34,10 +34,8 @@ public class PluginsTests : TestHelper
         collector.Expect("MyCompany.MyProduct.MyLibrary");
 #if NETFRAMEWORK
         collector.Expect("OpenTelemetry.Instrumentation.Http.HttpWebRequest", span => span.Attributes.Any(att => att.Key == "example.plugin"));
-#elif NET7_0_OR_GREATER
-        collector.Expect("System.Net.Http", span => span.Attributes.Any(att => att.Key == "example.plugin"));
 #else
-        collector.Expect("OpenTelemetry.Instrumentation.Http.HttpClient", span => span.Attributes.Any(att => att.Key == "example.plugin"));
+        collector.Expect("System.Net.Http", span => span.Attributes.Any(att => att.Key == "example.plugin"));
 #endif
 
         SetEnvironmentVariable("OTEL_DOTNET_AUTO_PLUGINS", "TestApplication.Plugins.Plugin, TestApplication.Plugins, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null");