+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Added
 
+- Support for .NET9.
 - Support for [RabbitMQ.Client](https://www.nuget.org/packages/RabbitMQ.Client/)
   traces instrumentation for versions `7.0.0`+.
 - Added bytecode instrumentation for `MongoDB` traces.
@@ -21,6 +22,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Removed
 
+- Support for .NET 6 and .NET 7. Both framework reached end of support.
 - Support for macOS Monterey 12 x64.
   macOs libraries are built and tested against [macOS Ventura 13 x64](https://github.com/actions/runner-images/blob/main/images/macos/macos-13-Readme.md).
 
@@ -38,6 +40,7 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
   - `OpenTelemetry.Instrumentation.Wcf` from `1.0.0-rc.17` to `1.0.0-rc.18`,
 - .NET Framework only, following packages updated
   - `Google.Protobuf` updated from `3.28.2` to `3.28.3`,
+  - `Grpc.Core.Api` from `2.66.0` to `2.67.0`,
   - `Microsoft.Extensions.DependencyInjection` from `8.0.0` to `8.0.1`,
   - `Microsoft.Extensions.DependencyInjection.Abstractions` from `8.0.1` to `8.0.2`,
   - `System.Diagnostics.DiagnosticSource` from `8.0.0` to `8.0.1`,